split orthology related components into individual files
