+++ src/js_src/components/orthology/index.js
@@ -1,0 +1,5 @@
+import OrthologyTable from './orthologyTable';
+
+export {
+  OrthologyTable
+};
